+++ lib/modules/test/test_service.dart
@@ -1,9 +1,9 @@
 ﻿// Module: test
-// Updated at: 2026-07-24 12:12:57
-// Commit: perf(test): apply RepaintBoundary on animated test widgets
+// Updated at: 2026-07-24 15:13:33
+// Commit: fix(test): fix memory leak by cancelling test subscription
 
 class testComponentState {
-  final String commitRef = '3103';
-  final String lastMessage = 'perf(test): apply RepaintBoundary on animated test widgets';
-  final DateTime timestamp = DateTime.parse('2026-07-24T12:12:57+05:00');
+  final String commitRef = '3116';
+  final String lastMessage = 'fix(test): fix memory leak by cancelling test subscription';
+  final DateTime timestamp = DateTime.parse('2026-07-24T15:13:33+05:00');
 }